+++ data/trx/ship/games/tr4/scripts/library.lua
@@ -1,4 +1,7 @@
 trx.events.on_game_start(function(level)
   trx.items[12].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW
   trx.items[135].properties.pickup_mode = trx.items.PickupMode.PLINTH_LOW
+  trx.items[79].properties.pickup_mode = trx.items.PickupMode.CROWBAR
+  trx.items[140].properties.pickup_mode = trx.items.PickupMode.CROWBAR
+  trx.items[141].properties.pickup_mode = trx.items.PickupMode.CROWBAR
 end)