+++ data/trx/ship/games/tr4/scripts/nutrench.lua
@@ -2,4 +2,5 @@ trx.events.on_game_start(function(level)
   trx.objects.animating_16.properties.collidable = false
   trx.objects.switch_type_generic_1.properties.switch_mode =
     trx.items.SwitchMode.SHOVE
+  trx.items[17].properties.pickup_mode = trx.items.PickupMode.CROWBAR
 end)