+++ src/trx/game/objects/general/pickup.h
@@ -9,6 +9,7 @@ typedef enum {
     PICKUP_MODE_PLINTH_LOW,
     PICKUP_MODE_PLINTH_HIGH,
     PICKUP_MODE_HIDDEN,
+    PICKUP_MODE_CROWBAR,
     PICKUP_MODE_NUMBER_OF,
 } PICKUP_MODE;
 
@@ -16,3 +17,4 @@ bool Pickup_Trigger(int16_t item_num);
 const OBJECT_BOUNDS *Pickup_Bounds(void);
 void Pickup_Collision(int16_t item_num, ITEM *lara_item, COLL_INFO *coll);
 uint32_t Pickup_GetSecretMask(const ITEM *item);
+int16_t Pickup_FindNearbyCrowbarPryPickup(void);